[ThinLTO] Change parameter type. NFC
Change destination module type for consistency with r345118

llvm-svn: 345124

+++ llvm/lib/IR/ModuleSummaryIndex.cpp
@@ -242,7 +242,7 @@ void ModuleSummaryIndex::exportToDot(raw_ostream& OS) const {
   };
 
   auto DrawEdge = [&](const char *Pfx, uint64_t SrcMod, GlobalValue::GUID SrcId,
-                      int DstMod, GlobalValue::GUID DstId, int TypeOrHotness) {
+                      uint64_t DstMod, GlobalValue::GUID DstId, int TypeOrHotness) {
     // 0 corresponds to alias edge, 1 to ref edge, 2 to call with unknown
     // hotness, ...
     TypeOrHotness += 2;
